Move where user defined paths are defined
